+++ javai-persistence/src/test/java/dev/xtrafe/javai/persistence/TestAssocTargetRepository.java
@@ -1,0 +1,4 @@
+package dev.xtrafe.javai.persistence;
+
+interface TestAssocTargetRepository extends JavAIRepository<TestAssocTarget> {
+}